+++ playwright/tests/keywordSearch.spec.ts
@@ -1,10 +1,6 @@
-import { Then, When } from "@cucumber/cucumber";
+import { Then } from "@cucumber/cucumber";
 import { expect } from "@playwright/test";
 import { CustomWorld } from "../support/setup";
-
-When("I input a search term", async function (this: CustomWorld) {
-  await this.page.locator('[aria-label="Item Search"]').fill("IBM 1401");
-});
 
 Then(
   "I count how many titles have my keyword",